more organization
Added a function called "selectForce". Now to add other types of forces they can be added anywhere as a function and be called trough that one. The loop in the main def for force calculations calls selectForce and selectForce selects the function (by only leaving the function you want to use uncommented)